+++ Src/Knowledge/AllSpellsPaladin.lua
@@ -19,7 +19,7 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :IsDefault(false)
                :Category("class")
 
-  local blessing_duration = allBuffsModule.TbcOrClassic(allBuffsModule.TEN_MINUTES, allBuffsModule.FIVE_MINUTES)
+  local blessingDuration = allBuffsModule.TbcOrClassic(allBuffsModule.TEN_MINUTES, allBuffsModule.FIVE_MINUTES)
   local greaterBlessingDuration = allBuffsModule.TbcOrClassic(allBuffsModule.FIFTEEN_MINUTES, allBuffsModule.HALF_AN_HOUR)
 
   --
@@ -30,10 +30,21 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :GroupFamily({ 25898 })
                :IsBlessing(true)
                :IsDefault(false)
-               :SingleDuration(blessing_duration)
+               :SingleDuration(blessingDuration)
                :DefaultTargetClasses({ "MAGE", "HUNTER", "WARLOCK" })
                :IgnoreIfHaveBuff(25898) -- Greater Kings
                :Category("blessing")
+               :HideInCata()
+
+  -- [Cataclysm] Blessing of Kings
+  buffDefModule:createAndRegisterBuff(allBuffs, 20217, nil)
+               :SingleFamily({ 20217, 79063 }) -- Cataclysm: Blessing of Kings (Spell id and actual aura id)
+               :GroupFamily({ 20217, 79063 }) -- Cataclysm: Blessing of Kings (Spell id and actual aura id)
+               :IsBlessing(true)
+               :IsDefault(false)
+               :SingleDuration(allBuffsModule.HOUR)
+               :Category("blessing")
+               :RequireCata()
 
   -- Blessing of Light (classic and TBC only)
   buffDefModule:createAndRegisterBuff(allBuffs, 19979, nil)
@@ -41,7 +52,7 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                                27144 }) -- TBC: Rank 4
                :IsBlessing(true)
                :IsDefault(false)
-               :SingleDuration(blessing_duration)
+               :SingleDuration(blessingDuration)
                :GroupDuration(greaterBlessingDuration)
                :RequirePlayerClass("PALADIN")
                :HideInWotLK()
@@ -56,23 +67,37 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :SingleFamily({ 19740, 19834, 19835, 19836, 19837, 19838, 25291, -- Blessing of Might 1-7
                                27140, -- TBC: Blessing of Might 8
                                48931, 48932 }) -- WotLK: Blessing of Might 9-10
-               :SingleDuration(blessing_duration)
+               :SingleDuration(blessingDuration)
                :RequirePlayerClass("PALADIN")
                :DefaultTargetClasses(allBuffsModule.PHYSICAL_CLASSES)
                :IgnoreIfHaveBuff({ 25782, 25916, 27141 }) -- Greater Might 1-3
                :IgnoreIfHaveBuff({ 48933, 48934 }) -- WotLK: Greater Might 4-5
                :Category("blessing")
+               :HideInCata()
+
+  -- [Cataclysm] Blessing of Might
+  buffDefModule:createAndRegisterBuff(allBuffs, 19740, nil)
+               :IsBlessing(true)
+               :IsDefault(true)
+               :SingleFamily({ 19740, 79102 }) -- Cataclysm: Blessing of Might (spell id and aura id)
+               :GroupFamily({ 19740, 79102 }) -- Cataclysm: Blessing of Might (spell id and aura id)
+               :SingleDuration(allBuffsModule.HOUR)
+               :RequirePlayerClass("PALADIN")
+               :DefaultTargetClasses(allBuffsModule.PHYSICAL_CLASSES)
+               :Category("blessing")
+               :RequireCata()
 
   -- Blessing of Salvation (classic and TBC only)
   buffDefModule:createAndRegisterBuff(allBuffs, 1038, nil)
                :IsBlessing(true)
                :IsDefault(false)
-               :SingleDuration(blessing_duration)
+               :SingleDuration(blessingDuration)
                :RequirePlayerClass("PALADIN")
                :HideInWotLK()
                :DefaultTargetClasses(allBuffsModule.BOM_NO_CLASSES)
                :IgnoreIfHaveBuff(25895) -- Greater Salv
                :Category("blessing")
+               :HideInCata()
 
   -- Blessing of Wisdom
   buffDefModule:createAndRegisterBuff(allBuffs, 25290, nil)
@@ -81,11 +106,12 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :SingleFamily({ 19742, 19850, 19852, 19853, 19854, 25290, -- Blessing of Wisdom 1-6
                                27142, -- TBC: Blessing of Wisdom 7
                                48935, 48936 }) -- WotLK: Blessing of Wisdom 8-9
-               :SingleDuration(blessing_duration)
+               :SingleDuration(blessingDuration)
                :GroupDuration(greaterBlessingDuration)
                :RequirePlayerClass("PALADIN")
                :DefaultTargetClasses(allBuffsModule.MANA_CLASSES)
                :Category("blessing")
+               :HideInCata()
 
   -- Blessing of Sanctuary
   buffDefModule:createAndRegisterBuff(allBuffs, 20914, nil)
@@ -93,7 +119,7 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :IsDefault(false)
                :SingleFamily({ 20911, 20912, 20913, 20914, -- Blessing of Sanctuary 1-4; WotLK: Blessing of Sanctuary
                                27168 }) -- TBC: Blessing of Sanctuary 5
-               :SingleDuration(blessing_duration)
+               :SingleDuration(blessingDuration)
                :GroupDuration(greaterBlessingDuration)
                :RequirePlayerClass("PALADIN")
                :DefaultTargetClasses(allBuffsModule.BOM_NO_CLASSES)
@@ -110,6 +136,7 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :RequirePlayerClass("PALADIN")
                :DefaultTargetClasses(allBuffsModule.BOM_NO_CLASSES)
                :Category("blessing")
+               :HideInCata()
 
   -- Greater Blessing of Light
   buffDefModule:createAndRegisterBuff(allBuffs, 25890, nil)
@@ -136,6 +163,7 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :RequirePlayerClass("PALADIN")
                :DefaultTargetClasses(allBuffsModule.PHYSICAL_CLASSES)
                :Category("blessing")
+               :HideInCata()
 
   --Greater Blessing of Salvation
   buffDefModule:createAndRegisterBuff(allBuffs, 25895, nil)
@@ -146,6 +174,7 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :RequirePlayerClass("PALADIN")
                :DefaultTargetClasses(allBuffsModule.BOM_NO_CLASSES)
                :Category("blessing")
+               :HideInCata()
 
   --Greater Blessing of Wisdom
   buffDefModule:createAndRegisterBuff(allBuffs, 25918, nil)
@@ -159,6 +188,7 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                :RequirePlayerClass("PALADIN")
                :DefaultTargetClasses(allBuffsModule.MANA_CLASSES)
                :Category("blessing")
+               :HideInCata()
 
   --Greater Blessing of Sanctuary
   buffDefModule:createAndRegisterBuff(allBuffs, 25899, nil)
@@ -285,6 +315,25 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                                27160 })  -- TBC: Seal of Light 5
                :RequirePlayerClass("PALADIN")
                :Category("seal")
+               :HideInCata()
+
+  -- [Cataclysm] Seal of Justice
+  buffDefModule:createAndRegisterBuff(allBuffs, 20164, nil)
+               :BuffType("seal")
+               :IsDefault(false)
+               :SingleFamily({ 20164 })  -- Cataclysm: Seal of Justice
+               :RequirePlayerClass("PALADIN")
+               :Category("seal")
+               :RequireCata()
+
+  -- [Cataclysm] Seal of Insight
+  buffDefModule:createAndRegisterBuff(allBuffs, 20165, nil)
+               :BuffType("seal")
+               :IsDefault(false)
+               :SingleFamily({ 20165 })  -- Cataclysm: Seal of Insight
+               :RequirePlayerClass("PALADIN")
+               :Category("seal")
+               :RequireCata()
 
   -- Seal of Righteousness
   buffDefModule:createAndRegisterBuff(allBuffs, 20154, nil)
@@ -295,6 +344,16 @@ function paladinModule:SetupPaladinSpells(allBuffs, enchants)
                                21084 }) -- WotLK: Seal of Righteousness
                :RequirePlayerClass("PALADIN")
                :Category("seal")
+               :HideInCata()
+
+  -- [Cataclysm] Seal of Righteousness
+  buffDefModule:createAndRegisterBuff(allBuffs, 20154, nil)
+               :BuffType("seal")
+               :IsDefault(false)
+               :SingleFamily({ 20154 }) -- Cataclysm: Seal of Righteousness
+               :RequirePlayerClass("PALADIN")
+               :Category("seal")
+               :RequireCata()
 
   -- Seal of Wisdom
   buffDefModule:createAndRegisterBuff(allBuffs, 20166, nil)