+++ Src/KvLib/KvEnv.lua
@@ -4,6 +4,8 @@
 ---@field haveTBC boolean
 ---@field isWotLK boolean
 ---@field haveWotLK boolean
+---@field isCata boolean
+---@field haveCata boolean
 local envModule = { }
 
 ---@class KvModuleManager
@@ -16,11 +18,14 @@ KvModuleManager = {
 
 function envModule:DetectVersions()
   local _, _, _, tocversion = GetBuildInfo()
-  self.isWotLK = (tocversion >= 30000 and tocversion <= 39999) -- TODO: change to WOTLK detection via WOW_PROJECT_..._CLASSIC
-  self.haveWotLK = self.isWotLK
+  self.isCata = WOW_PROJECT_ID == WOW_PROJECT_CATACLYSM_CLASSIC
+  self.haveCata = self.isCata
+
+  self.isWotLK = WOW_PROJECT_ID == WOW_PROJECT_WRATH_CLASSIC
+  self.haveWotLK = self.isWotLK or self.isCata
 
   self.isTBC = WOW_PROJECT_ID == WOW_PROJECT_BURNING_CRUSADE_CLASSIC
-  self.haveTBC = self.isWotLK or self.isTBC
+  self.haveTBC = self.isWotLK or self.isTBC or self.isCata
 
   self.isClassic = WOW_PROJECT_ID == WOW_PROJECT_CLASSIC